+++ plugins/trivy/pkg/image/util_test.go
@@ -1,6 +1,7 @@
 package image
 
 import (
+	"os"
 	"testing"
 
 	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/models"
@@ -65,4 +66,36 @@ func TestToScanMatches(t *testing.T) {
 	matching = GetUnmatchingImages(getOrigReportForTest(), toScan, false)
 	assert.Equal(t, 3, len(matching))
 	assert.Equal(t, "quay.io/fairwinds/sample-1:1.2.3", matching[0].Name)
+}
+
+func TestShouldBeAbleToReadOldReports(t *testing.T) {
+	v1Body, err := os.ReadFile("testdata/v0.26/latest.json")
+	assert.NoError(t, err)
+
+	v2, err := unmarshalAndFixReport(v1Body)
+	assert.NoError(t, err)
+	assert.Equal(t, 28, len(v2.Images))
+	assert.Equal(t, 467, len(v2.Vulnerabilities))
+
+	for _, img := range v2.Images {
+		if img.RecommendationOnly {
+			assert.Len(t, img.Owners, 0)
+		} else {
+			assert.Len(t, img.Owners, 1)
+		}
+	}
+}
+
+func TestUnmarshalAndFixReport(t *testing.T) {
+	v2Body, err := os.ReadFile("testdata/v0.27/latest.json")
+	assert.NoError(t, err)
+
+	v2, err := unmarshalAndFixReport(v2Body)
+	assert.NoError(t, err)
+	assert.Equal(t, 3, len(v2.Images))
+	assert.Equal(t, 467, len(v2.Vulnerabilities))
+
+	assert.Len(t, v2.Images[0].Owners, 1)
+	assert.Len(t, v2.Images[1].Owners, 1)
+	assert.Len(t, v2.Images[2].Owners, 2)
 }
