+++ plugins/trivy/pkg/util/util.go
@@ -1,13 +1,11 @@
 package util
 
 import (
-	"encoding/json"
 	"errors"
 	"os"
 	"os/exec"
 	"strings"
 
-	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/models"
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/client-go/dynamic"
@@ -19,7 +17,6 @@ import (
 )
 
 const UnknownOSMessage = "Unknown OS"
-const DockerIOprefix = "docker.io/"
 
 // KubeClientResources bundles together Kubernetes clients and related
 // resources.
@@ -90,49 +87,4 @@ func RunCommand(cmd *exec.Cmd, message string) error {
 		}
 	}
 	return err
-}
-
-func UnmarshalAndFixReport(body []byte) (*models.MinimizedReport, error) {
-	var report models.MinimizedReport
-	err := json.Unmarshal(body, &report)
-	if err != nil {
-		return nil, err
-	}
-	fixOwners(&report)
-	normalizeDockerHubImages(&report)
-	return &report, nil
-}
-
-// fixOwners adapt older owners fields to the new ones
-func fixOwners(report *models.MinimizedReport) {
-	for i := range report.Images {
-		img := &report.Images[i]
-		if hasDeprecatedOwnerFields(*img) {
-			var container string
-			if img.OwnerContainer != nil {
-				container = *img.OwnerContainer
-			}
-			img.Owners = []models.Resource{
-				{
-					Name:      img.OwnerName,
-					Kind:      img.OwnerKind,
-					Namespace: img.Namespace,
-					Container: container,
-				},
-			}
-		}
-	}
-}
-
-func hasDeprecatedOwnerFields(img models.ImageDetailsWithRefs) bool {
-	return len(img.OwnerName) != 0 || len(img.OwnerKind) != 0 || len(img.Namespace) != 0
-}
-
-// normalizeDockerHubImages removes the docker.io/ prefix from the image names and IDs
-func normalizeDockerHubImages(report *models.MinimizedReport) {
-	for i := range report.Images {
-		img := &report.Images[i]
-		img.Name = strings.TrimPrefix(img.Name, DockerIOprefix)
-		img.ID = strings.TrimPrefix(img.ID, DockerIOprefix)
-	}
 }
